# music/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music API",
        "description": "Documentation de l'API Musicbox",
        "contact": {
            "name": "Ayaah Mhaouar",
            "email": "ayaah@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/playlists": {
            "get": {
                "tags": [
                    "Playlist"
                ],
                "summary": "Lister toutes les playlists",
                "operationId": "758aa7244638ce70d4ce78a5c27ac3a5",
                "responses": {
                    "200": {
                        "description": "Liste de playlists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "My Playlist"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Playlist"
                ],
                "summary": "Créer une playlist",
                "operationId": "87ac9a92f0fb81a38eeaf7699cbd0e86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Workout Mix"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Playlist créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Workout Mix"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/playlists/{id}": {
            "get": {
                "tags": [
                    "Playlist"
                ],
                "summary": "Afficher une playlist avec ses musiques",
                "operationId": "6cbaf5a7258a2c09b653716d841b9ad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Playlist trouvée avec ses musiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Chill Vibes"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "musics": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Lo-Fi Beat"
                                                    },
                                                    "artist": {
                                                        "type": "string",
                                                        "example": "DJ Relax"
                                                    },
                                                    "album": {
                                                        "type": "string",
                                                        "example": "Lo-Fi Collection"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 180
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Playlist"
                ],
                "summary": "Modifier une playlist",
                "operationId": "c2d51e511d6cf4cc556481f19316402d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Playlist"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Playlist mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Playlist"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Playlist"
                ],
                "summary": "Supprimer une playlist",
                "operationId": "8d5370dcdb5fde9a0ebf49aa2a5f563c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Playlist supprimée"
                    }
                }
            }
        },
...
    },
...
}